+++ crates/client/rpc/src/methods/read/estimate_message_fee.rs
@@ -1,13 +1,15 @@
 use std::sync::Arc;
 
 use blockifier::transaction::transactions::L1HandlerTransaction;
-use dp_felt::{Felt252Wrapper, FeltWrapper};
+use dp_convert::felt_wrapper::FeltWrapper;
 use dp_transactions::compute_hash::ComputeTransactionHash;
 use jsonrpsee::core::RpcResult;
 use starknet_api::core::Nonce;
+use starknet_api::core::{ContractAddress, EntryPointSelector};
 use starknet_api::hash::StarkFelt;
 use starknet_api::transaction::{Calldata, Fee, TransactionVersion};
 use starknet_core::types::{BlockId, FeeEstimate, MsgFromL1};
+use starknet_types_core::felt::Felt;
 
 use crate::utils::execution::block_context;
 use crate::utils::ResultExt;
@@ -38,7 +40,8 @@ pub async fn estimate_message_fee(
     let block_context = block_context(starknet, &block_info)?;
     let block_number = block_info.block_n();
 
-    let transaction = convert_message_into_tx(message, starknet.chain_config.chain_id.0.into(), Some(block_number));
+    let chain_id = starknet.chain_config.chain_id;
+    let transaction = convert_message_into_tx(message, chain_id, Some(block_number));
 
     let message_fee = utils::execution::estimate_message_fee(starknet, transaction, &block_context)
         .or_contract_error("Function execution failed")?;
@@ -46,19 +49,17 @@ pub async fn estimate_message_fee(
     Ok(message_fee)
 }
 
-pub fn convert_message_into_tx(
-    message: MsgFromL1,
-    chain_id: Felt252Wrapper,
-    block_number: Option<u64>,
-) -> L1HandlerTransaction {
+pub fn convert_message_into_tx(message: MsgFromL1, chain_id: Felt, block_number: Option<u64>) -> L1HandlerTransaction {
     let calldata = std::iter::once(message.from_address.into_stark_felt())
         .chain(message.payload.into_iter().map(FeltWrapper::into_stark_felt))
         .collect();
+
     let tx = starknet_api::transaction::L1HandlerTransaction {
         version: TransactionVersion::ZERO,
         nonce: Nonce(StarkFelt::ZERO),
-        contract_address: Felt252Wrapper::from(message.to_address).into(),
-        entry_point_selector: Felt252Wrapper::from(message.entry_point_selector).into(),
+        contract_address: ContractAddress::try_from(message.to_address.into_stark_felt())
+            .expect("expected a contact address"),
+        entry_point_selector: EntryPointSelector(message.entry_point_selector.into_stark_felt()),
         calldata: Calldata(Arc::new(calldata)),
     };
     // TODO(merge): recheck if this is correct