+++ crates/client/rpc/src/methods/read/get_transaction_receipt.rs
@@ -2,7 +2,7 @@ use blockifier::context::BlockContext;
 use blockifier::transaction::objects::TransactionExecutionInfo;
 use blockifier::transaction::transaction_execution as btx;
 use dc_db::mapping_db::BlockStorageType;
-use dp_felt::FeltWrapper;
+use dp_convert::felt_wrapper::FeltWrapper;
 use jsonrpsee::core::RpcResult;
 use starknet_api::core::{calculate_contract_address, ContractAddress};
 use starknet_api::transaction::{Transaction, TransactionHash};