+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
 - fix(logs): fixed logs and get_state_update
+- refactor: remove primitives/felt
+- refactor: move convert.rs to felt_wrapper.rs
 - fix(decode): fix mapping db decoding
 - feat: store reverted txs hashes
 - feat(l1): added l1 free rpc url if none is provided